initial commit
motivation:

1. create tooling to make using qutebrowser sessions for more purposes
   easier
2. learn to use htmx so I can compellingly convince people to stop
   having separate codebases for their frontends and backends

use cases:

* move accumulated tabs from my main session into session representing
  categories - probably a bookmark/link manager would be better for this
* better UI to rearrange the tabs in my main session

I'm using flask and html because I'm familiar with flask and I though
using this stack would be a smaller workload than brushing up on some
TUI option. I've already had to add in some JS to do drag and drop
though, which doesn't quite line up with my initial UI for selecting
tabs (the multi select doesn't work).

There's quite a few todos left to pick from, but I think almost all the
the functionality to make it useful is there. Although for some of it I
might have to teach it about my out of tree lazy loading state file.

I've been focussing on adding in features to get enough code to refactor
and I think it's about large enough to be unwieldy now, so it's about
time.