Filter out empty windows when saving
I was running into an issue when moving tabs between sessions. If I
moved all of the tabs out of a window, that window would be left empty.
But I'm hiding empty windows when rendering the template. That was
causing two issues:
1. when moving tabs between windows, if there was an empty window in the
   session that tabs were moving from, then the window indices would be
   offset compared to the raw data. This could cause an issue when the
   code went to move the tab out of the wrong window
2. when moving tabs to a session they are always moved to the last
   window of the session. If the window wasn't visible in the UI this
   could be surprising.

To make things work as I expect I've flowed the raw window indexes
through to the template, and filtered out empty windows when saving to
get rid of them.

I'm not sure this is the best fix, there are a few missing features
around windows. We can't create new ones, can't delete them, can't move
items from one sessions to an arbitrary tab in a different one.

Probably if there was a way to create and delete windows I could just
always show the empty ones, and stop filtering them when saving. Being
able to create windows seems useful for being able to slice and dice
sessions anyhow.

Also saved the "from session" when moving tabs. Probably an oversight,
but I feel like I left that out to avoid having to copy default back for
more testing? Maybe.